Invocation queue is the new Worker
